Change IGuiPosition + IGuiArea to use double's rather than int's. Many bugs exist now.

+++ common/buildcraft/lib/client/sprite/SpriteNineSliced.java
@@ -44,7 +44,7 @@ public class SpriteNineSliced {
         draw(element.getX(), element.getY(), element.getWidth(), element.getHeight());
     }
 
-    public void draw(int x, int y, int width, int height) {
+    public void draw(double x, double y, double width, double height) {
         sprite.bindTexture();
         Tessellator tess = Tessellator.getInstance();
         VertexBuffer vb = tess.getBuffer();
